backend/requirements.txt
amqp==2.6.1
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.3.0
celery==4.4.1
Django==3.1.6
djangorestframework==3.12.2
drf-spectacular==0.13.2
importlib-metadata==3.4.0
inflection==0.5.1
jsonschema==3.2.0
kombu==4.6.11
Pillow==8.1.0
psycopg2==2.8.6
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
redis==3.4.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
vine==1.3.0
zipp==3.4.0
django-cors-headers==3.7.0
